(documentation change) Specify full path to /usr/bin/xattr for macOS, fixes #2583
On some Macs, Homebrew installs its own `xattr` which doesn't have the `-r` option. See #2583.

This PR just adds the full path to the system `xattr` in `README.md`, so the `-r` in the instructions doesn't cause a problem.

It's not clear to me whether or not the `-r` is required; this PR does not address that question, it just routes around it.

I have tested the instruction on my Mac which has the Homebrew `xattr`, and with the added path, it works.

I would assume this change is low risk; it's just to documentation, and the change points to a system command that should always be there. However, it's possible I'm missing corner cases that further testing and validation on more machines would uncover.